Depreciated legacy event management components, added new event manager interface
