+++ Runtime/EventManager.cs.meta
@@ -1,3 +1,11 @@
 fileFormatVersion: 2
-guid: 78da829307d6437cbf7e0caf41b01d8b
-timeCreated: 1688360835
+guid: 25d441645c4bef744bfb7ab853f2884f
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
