+++ Runtime/Legacy/IEventManager.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 0e1bbd2119d682d4c9de051901592642
+guid: 20a42312bc1017d478fec633d842781d
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
