+++ Tests/Runtime/Legacy.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 85110606641bcbb4b9803efe7bc55a5e
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
