+++ Tests/Runtime/Legacy/GlobalEventManagerTests.cs.meta
